hexlet-code/pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
main="charts_project.main:run_all"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
